requirements.txt
bcrypt==3.2.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
markdown==3.6.0
mysql-connector-python==8.0.28
python-decouple==3.5
requests==2.26.0
Werkzeug==2.2.2
importlib-metadata==4.13.0
kafka-python==2.0.2
